refactor: split QuestionItem into per-type components
