+++ src/components/student/QuestionItem/index.ts
@@ -1,0 +1,1 @@
+export { QuestionItem } from "./QuestionItem";
